refactor & fix: page into components and fixed global.css bug
