+++ src/app.html
@@ -5,7 +5,7 @@
 		<meta name="description" content="TuMi Mundo - De app om je kind talen te leren.">
 		<meta name="robots" content="noindex, nofollow">
 		<link rel="icon" href="%sveltekit.assets%/favicon.png" />
-        <link rel="stylesheet" href="%sveltekit.assets%/style.css" />
+        <link rel="stylesheet" href="%sveltekit.assets%/global.css" />
 		<link rel="preconnect" href="https://fonts.googleapis.com">
 		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 		<link href="https://fonts.googleapis.com/css2?family=Poppins&display=swap" rel="stylesheet">
