+++ src/lib/index.js
@@ -20,6 +20,9 @@ export { default as SuggestedPlaylists } from '$lib/components/core/SuggestedPla
 export { default as LoginIntro } from '$lib/components/core/LoginIntro.svelte'
 export { default as LoginPopup } from '$lib/components/core/LoginPopup.svelte'
 
+export { default as SignUpForm } from '$lib/components/core/SignUpForm.svelte'
+
+
 
 
 
@@ -29,12 +32,16 @@ export { default as Input } from '$lib/components/forms/Input.svelte'
 export { default as Search } from '$lib/components/forms/Search.svelte'
 export { default as AddStory } from '$lib/components/forms/AddStory.svelte';
 export { default as CarouselIndicators } from '$lib/components/forms/CarouselIndicators.svelte'
+export { default as PasswordInput } from '$lib/components/forms/PasswordInput.svelte'
 
 export { default as Playlist } from '$lib/components/layout/Playlist.svelte'
 export { default as Popup } from '$lib/components/layout/Popup.svelte'
 export { default as Story } from '$lib/components/layout/Story.svelte'
 
 export { default as OnboardingSlide } from '$lib/components/layout/OnboardingSlide.svelte'
+export { default as FormHeading } from '$lib/components/layout/FormHeading.svelte'
+export { default as TermsToggles } from '$lib/components/layout/TermsToggles.svelte'
+
 
 
 // Helpers
